==> preprocessing/00-visualize.ipynb <==
import pandas as pd
from sklearn.preprocessing import OneHotEncoder
import matplotlib.pyplot as plt
import plotly.express as px
import seaborn as sns
import numpy as np
import ast

pd.set_option("display.max_columns", None)

df = pd.read_csv('../data/train.csv')
df.head()

df.shape

df.describe()

print(df.columns)

df.describe()

# plotting target variable

sns.histplot(df['price'], bins=50)
plt.title('Distribution of Airbnb Prices')
plt.xlabel('Price')
plt.ylabel('Count')
plt.show()

# plotting the airbnb listings by location and bed price

df.dropna(subset=['price', 'latitude', 'longitude', 'neighbourhood_cleansed', 'beds'], inplace=True)

# include the neighborhood, price, and number of beds in the hover data
fig = px.scatter_mapbox(df, lat="latitude", lon="longitude", hover_name="price", hover_data=["neighbourhood_cleansed", "neighbourhood_group_cleansed", "accommodates"],
                        color="price", 
                        color_continuous_scale=px.colors.sequential.Jet, 
                        size_max=30, zoom=9.5, height=600)
fig.update_layout(mapbox_style="open-street-map")
fig.update_layout(margin={"r":0,"t":0,"l":0,"b":0})

# plot some landmarks in NYC
landmarks = pd.DataFrame({
    'name': ['Statue of Liberty', 'Central Park', 'Empire State Building', 'Brooklyn Bridge'],
    'latitude': [40.6892, 40.785091, 40.748817, 40.706086],
    'longitude': [-74.0445, -73.968285, -73.985428, -73.996864]
})
landmarks = pd.concat([landmarks, pd.DataFrame({
    'name': ['Times Square', 'Rockefeller Center', 'One World Trade Center', 'Broadway Theatre'],
    'latitude': [40.7580, 40.7587, 40.7127, 40.7590],
    'longitude': [-73.9855, -73.9787, -74.0134, -73.9845]
})])

for i, row in landmarks.iterrows():
    fig.add_scattermapbox(
        lat=[row['latitude']],
        lon=[row['longitude']],
        mode='markers+text',
        marker=dict(size=15, color='white'),
        text=row['name'],
        textposition='top center',
        showlegend=False
    )

fig.show()

# plot all distributions of the numerical features
df.hist(bins=50, figsize=(30, 15))
