==> preprocessing/04-combine.ipynb <==
import pandas as pd

# combining the results of the review sentiment analysis and the previous feature engineering
train_nlp1 = pd.read_csv('../data/train_nlp1.csv')
train_sentiment = pd.read_csv('../data/train_sentiment.csv')

train = pd.concat([train_nlp1, train_sentiment], axis=1)

train.drop(columns=['reviews'], inplace=True)


train.loc[train['number_of_reviews'] == 0, 'avg_review_length'] = None

train.to_csv('../data/train_nlp2.csv', index=False)


test_nlp1 = pd.read_csv('../data/test_nlp1.csv')
test_sentiment = pd.read_csv('../data/test_sentiment.csv')

test = pd.concat([test_nlp1, test_sentiment], axis=1)

test.drop(columns=['reviews'], inplace=True)

test.loc[test['number_of_reviews'] == 0, 'avg_review_length'] = None

test.to_csv('../data/test_nlp2.csv', index=False)
